refactor: Break up editingV2Project file
