+++ src/redux/slices/editingV2Project/defaultState.ts
@@ -1,0 +1,144 @@
+import { BigNumber } from '@ethersproject/bignumber'
+import * as constants from '@ethersproject/constants'
+import { JB721_DELEGATE_V1 } from 'constants/delegateVersions'
+import { ONE_MILLION } from 'constants/numbers'
+import {
+  ETH_PAYOUT_SPLIT_GROUP,
+  RESERVED_TOKEN_SPLIT_GROUP,
+} from 'constants/splits'
+import { JB721GovernanceType, JBTiered721Flags } from 'models/nftRewardTier'
+import {
+  LATEST_METADATA_VERSION,
+  ProjectMetadataV5,
+} from 'models/project-metadata'
+import { issuanceRateFrom, redemptionRateFrom } from 'utils/v2v3/math'
+import {
+  SerializedV2V3FundingCycleData,
+  SerializedV2V3FundingCycleMetadata,
+  serializeV2V3FundingCycleData,
+  serializeV2V3FundingCycleMetadata,
+} from 'utils/v2v3/serializers'
+import { CreateState, ProjectState, ReduxState } from './types'
+
+export const DEFAULT_MUST_START_AT_OR_AFTER = '1'
+export const DEFAULT_MINT_RATE = ONE_MILLION
+
+// Increment this version by 1 when making breaking or major changes.
+// When users return to the site and their local version is less than
+// this number, their state will be reset.
+export const REDUX_STORE_V2_PROJECT_VERSION = 12
+
+// IF CHANGING ANY DEFAULT STATE, MUST INCREMENT REDUX VERSION ABOVE (`REDUX_STORE_V2_PROJECT_VERSION`)
+export const DEFAULT_FUNDING_CYCLE_DATA: SerializedV2V3FundingCycleData =
+  serializeV2V3FundingCycleData({
+    duration: BigNumber.from(0),
+    weight: BigNumber.from(issuanceRateFrom(DEFAULT_MINT_RATE.toString())), // 1e24, resulting in 1,000,000 tokens per ETH
+    discountRate: BigNumber.from(0), // A number from 0-1,000,000,000
+    ballot: constants.AddressZero,
+  })
+
+// IF CHANGING ANY DEFAULT STATE, MUST INCREMENT REDUX VERSION ABOVE (`REDUX_STORE_V2_PROJECT_VERSION`)
+export const DEFAULT_FUNDING_CYCLE_METADATA: SerializedV2V3FundingCycleMetadata =
+  serializeV2V3FundingCycleMetadata({
+    global: {
+      allowSetTerminals: false,
+      allowSetController: false,
+      pauseTransfers: false,
+    },
+    reservedRate: BigNumber.from(0), // A number from 0-10,000
+    redemptionRate: redemptionRateFrom('100'), // A number from 0-10,000
+    ballotRedemptionRate: redemptionRateFrom('100'), // A number from 0-10,000
+    pausePay: false,
+    pauseDistributions: false,
+    pauseRedeem: false,
+    allowMinting: false,
+    pauseBurn: false,
+    preferClaimedTokenOverride: false,
+    allowTerminalMigration: false,
+    allowControllerMigration: false,
+    holdFees: false,
+    useTotalOverflowForRedemptions: false,
+    useDataSourceForPay: false,
+    useDataSourceForRedeem: false,
+    dataSource: constants.AddressZero,
+    metadata: BigNumber.from(0),
+  }) ?? {}
+
+// IF CHANGING ANY DEFAULT STATE, MUST INCREMENT REDUX VERSION ABOVE (`REDUX_STORE_V2_PROJECT_VERSION`)
+export const EMPTY_PAYOUT_GROUPED_SPLITS = {
+  group: ETH_PAYOUT_SPLIT_GROUP,
+  splits: [],
+}
+
+// IF CHANGING ANY DEFAULT STATE, MUST INCREMENT REDUX VERSION ABOVE (`REDUX_STORE_V2_PROJECT_VERSION`)
+export const EMPTY_RESERVED_TOKENS_GROUPED_SPLITS = {
+  group: RESERVED_TOKEN_SPLIT_GROUP,
+  splits: [],
+}
+
+// IF CHANGING ANY DEFAULT STATE, MUST INCREMENT REDUX VERSION ABOVE (`REDUX_STORE_V2_PROJECT_VERSION`)
+export const EMPTY_NFT_COLLECTION_METADATA = {
+  symbol: undefined,
+  name: undefined,
+  CID: undefined,
+  description: undefined,
+}
+
+// IF CHANGING ANY DEFAULT STATE, MUST INCREMENT REDUX VERSION ABOVE (`REDUX_STORE_V2_PROJECT_VERSION`)
+export const DEFAULT_NFT_FLAGS: JBTiered721Flags = {
+  lockReservedTokenChanges: false,
+  lockVotingUnitChanges: false,
+  lockManualMintingChanges: false,
+  preventOverspending: false,
+}
+
+// IF CHANGING ANY DEFAULT STATE, MUST INCREMENT REDUX VERSION ABOVE (`REDUX_STORE_V2_PROJECT_VERSION`)
+const DEFAULT_PROJECT_METADATA_STATE: ProjectMetadataV5 = {
+  name: '',
+  infoUri: '',
+  logoUri: '',
+  description: '',
+  twitter: '',
+  discord: '',
+  tokens: [],
+  nftPaymentSuccessModal: undefined,
+  version: LATEST_METADATA_VERSION,
+}
+
+// IF CHANGING ANY DEFAULT STATE, MUST INCREMENT REDUX VERSION ABOVE (`REDUX_STORE_V2_PROJECT_VERSION`)
+const DEFAULT_CREATE_STATE: CreateState = {
+  reconfigurationRuleSelection: undefined,
+  fundingCyclesPageSelection: undefined,
+  createFurthestPageReached: 'projectDetails',
+  createSoftLockPageQueue: [],
+  fundingTargetSelection: undefined,
+  payoutsSelection: undefined,
+  projectTokensSelection: undefined,
+}
+
+// IF CHANGING ANY DEFAULT STATE, MUST INCREMENT REDUX VERSION ABOVE (`REDUX_STORE_V2_PROJECT_VERSION`)
+const DEFAULT_PROJECT_STATE: ProjectState = {
+  projectMetadata: { ...DEFAULT_PROJECT_METADATA_STATE },
+  fundingCycleData: { ...DEFAULT_FUNDING_CYCLE_DATA },
+  fundingCycleMetadata: { ...DEFAULT_FUNDING_CYCLE_METADATA },
+  fundAccessConstraints: [],
+  payoutGroupedSplits: EMPTY_PAYOUT_GROUPED_SPLITS,
+  reservedTokensGroupedSplits: EMPTY_RESERVED_TOKENS_GROUPED_SPLITS,
+  nftRewards: {
+    rewardTiers: [],
+    CIDs: undefined,
+    collectionMetadata: EMPTY_NFT_COLLECTION_METADATA,
+    postPayModal: undefined,
+    flags: DEFAULT_NFT_FLAGS,
+    contractVersion: JB721_DELEGATE_V1,
+    governanceType: JB721GovernanceType.NONE,
+  },
+  mustStartAtOrAfter: DEFAULT_MUST_START_AT_OR_AFTER,
+  inputProjectOwner: undefined,
+}
+
+export const DEFAULT_REDUX_STATE: ReduxState = {
+  version: REDUX_STORE_V2_PROJECT_VERSION,
+  ...DEFAULT_PROJECT_STATE,
+  ...DEFAULT_CREATE_STATE,
+}
