+++ src/redux/slices/editingV2Project/editingV2Project.ts
@@ -1,214 +1,37 @@
-import { BigNumber } from '@ethersproject/bignumber'
-import * as constants from '@ethersproject/constants'
 import { createSlice, PayloadAction } from '@reduxjs/toolkit'
 import { AllocationSplit } from 'components/Allocation'
 import { projectTokenSettingsToReduxFormat } from 'components/Create/utils/projectTokenSettingsToReduxFormat'
-import { JB721_DELEGATE_V1 } from 'constants/delegateVersions'
-import {
-  ETH_PAYOUT_SPLIT_GROUP,
-  RESERVED_TOKEN_SPLIT_GROUP,
-} from 'constants/splits'
 import { CreatePage } from 'models/create-page'
-import { FundingTargetType } from 'models/fundingTargetType'
 import {
-  JBTiered721Flags,
-  DelegateVersion,
   JB721GovernanceType,
   NftCollectionMetadata,
   NftPostPayModalConfig,
   NftRewardTier,
 } from 'models/nftRewardTier'
 import { PayoutsSelection } from 'models/payoutsSelection'
-import {
-  LATEST_METADATA_VERSION,
-  ProjectMetadataV5,
-} from 'models/project-metadata'
 import { ProjectTokensSelection } from 'models/projectTokenSelection'
 import { ReconfigurationStrategy } from 'models/reconfigurationStrategy'
-import {
-  ETHPayoutGroupedSplits,
-  ReservedTokensGroupedSplits,
-  Split,
-} from 'models/splits'
-import {
-  DEFAULT_MINT_RATE,
-  issuanceRateFrom,
-  redemptionRateFrom,
-} from 'utils/v2v3/math'
+import { Split } from 'models/splits'
 import {
   SerializedV2V3FundAccessConstraint,
   SerializedV2V3FundingCycleData,
   SerializedV2V3FundingCycleMetadata,
-  serializeV2V3FundingCycleData,
-  serializeV2V3FundingCycleMetadata,
 } from 'utils/v2v3/serializers'
-
-export type NftRewardsData = {
-  rewardTiers: NftRewardTier[] | undefined
-  CIDs: string[] | undefined // points to locations of the NFTs' json on IPFS
-  collectionMetadata: NftCollectionMetadata
-  postPayModal: NftPostPayModalConfig | undefined
-  flags: JBTiered721Flags
-  contractVersion: DelegateVersion | undefined
-  governanceType: JB721GovernanceType
-}
-
-export interface CreateState {
-  fundingCyclesPageSelection: 'automated' | 'manual' | undefined
-  fundingTargetSelection: FundingTargetType | undefined
-  payoutsSelection: PayoutsSelection | undefined
-  projectTokensSelection: ProjectTokensSelection | undefined
-  reconfigurationRuleSelection: ReconfigurationStrategy | undefined
-  createFurthestPageReached: CreatePage
-  createSoftLockPageQueue: CreatePage[] | undefined
-}
-
-export interface ProjectState {
-  projectMetadata: ProjectMetadataV5
-  fundingCycleData: SerializedV2V3FundingCycleData
-  fundingCycleMetadata: SerializedV2V3FundingCycleMetadata
-  fundAccessConstraints: SerializedV2V3FundAccessConstraint[]
-  payoutGroupedSplits: ETHPayoutGroupedSplits
-  reservedTokensGroupedSplits: ReservedTokensGroupedSplits
-  nftRewards: NftRewardsData
-  mustStartAtOrAfter: string
-  inputProjectOwner: string | undefined
-}
-
-interface ReduxState extends CreateState, ProjectState {
-  version: number
-}
-
-// Increment this version by 1 when making breaking or major changes.
-// When users return to the site and their local version is less than
-// this number, their state will be reset.
-export const REDUX_STORE_V2_PROJECT_VERSION = 12
-
-export const DEFAULT_MUST_START_AT_OR_AFTER = '1'
-
-// IF CHANGING ANY DEFAULT STATE, MUST INCREMENT REDUX VERSION ABOVE (`REDUX_STORE_V2_PROJECT_VERSION`)
-export const defaultFundingCycleData: SerializedV2V3FundingCycleData =
-  serializeV2V3FundingCycleData({
-    duration: BigNumber.from(0),
-    weight: BigNumber.from(issuanceRateFrom(DEFAULT_MINT_RATE.toString())), // 1e24, resulting in 1,000,000 tokens per ETH
-    discountRate: BigNumber.from(0), // A number from 0-1,000,000,000
-    ballot: constants.AddressZero,
-  })
-
-// IF CHANGING ANY DEFAULT STATE, MUST INCREMENT REDUX VERSION ABOVE (`REDUX_STORE_V2_PROJECT_VERSION`)
-export const defaultFundingCycleMetadata: SerializedV2V3FundingCycleMetadata =
-  serializeV2V3FundingCycleMetadata({
-    global: {
-      allowSetTerminals: false,
-      allowSetController: false,
-      pauseTransfers: false,
-    },
-    reservedRate: BigNumber.from(0), // A number from 0-10,000
-    redemptionRate: redemptionRateFrom('100'), // A number from 0-10,000
-    ballotRedemptionRate: redemptionRateFrom('100'), // A number from 0-10,000
-    pausePay: false,
-    pauseDistributions: false,
-    pauseRedeem: false,
-    allowMinting: false,
-    pauseBurn: false,
-    preferClaimedTokenOverride: false,
-    allowTerminalMigration: false,
-    allowControllerMigration: false,
-    holdFees: false,
-    useTotalOverflowForRedemptions: false,
-    useDataSourceForPay: false,
-    useDataSourceForRedeem: false,
-    dataSource: constants.AddressZero,
-    metadata: BigNumber.from(0),
-  }) ?? {}
-
-// IF CHANGING ANY DEFAULT STATE, MUST INCREMENT REDUX VERSION ABOVE (`REDUX_STORE_V2_PROJECT_VERSION`)
-const EMPTY_PAYOUT_GROUPED_SPLITS = {
-  group: ETH_PAYOUT_SPLIT_GROUP,
-  splits: [],
-}
-
-// IF CHANGING ANY DEFAULT STATE, MUST INCREMENT REDUX VERSION ABOVE (`REDUX_STORE_V2_PROJECT_VERSION`)
-export const EMPTY_RESERVED_TOKENS_GROUPED_SPLITS = {
-  group: RESERVED_TOKEN_SPLIT_GROUP,
-  splits: [],
-}
-
-// IF CHANGING ANY DEFAULT STATE, MUST INCREMENT REDUX VERSION ABOVE (`REDUX_STORE_V2_PROJECT_VERSION`)
-export const EMPTY_NFT_COLLECTION_METADATA = {
-  symbol: undefined,
-  name: undefined,
-  CID: undefined,
-  description: undefined,
-}
-
-// IF CHANGING ANY DEFAULT STATE, MUST INCREMENT REDUX VERSION ABOVE (`REDUX_STORE_V2_PROJECT_VERSION`)
-export const DEFAULT_NFT_FLAGS: JBTiered721Flags = {
-  lockReservedTokenChanges: false,
-  lockVotingUnitChanges: false,
-  lockManualMintingChanges: false,
-  preventOverspending: false,
-}
-
-// IF CHANGING ANY DEFAULT STATE, MUST INCREMENT REDUX VERSION ABOVE (`REDUX_STORE_V2_PROJECT_VERSION`)
-const defaultCreateState: CreateState = {
-  reconfigurationRuleSelection: undefined,
-  fundingCyclesPageSelection: undefined,
-  createFurthestPageReached: 'projectDetails',
-  createSoftLockPageQueue: [],
-  fundingTargetSelection: undefined,
-  payoutsSelection: undefined,
-  projectTokensSelection: undefined,
-}
-
-// IF CHANGING ANY DEFAULT STATE, MUST INCREMENT REDUX VERSION ABOVE (`REDUX_STORE_V2_PROJECT_VERSION`)
-const defaultProjectMetadataState: ProjectMetadataV5 = {
-  name: '',
-  infoUri: '',
-  logoUri: '',
-  description: '',
-  twitter: '',
-  discord: '',
-  tokens: [],
-  nftPaymentSuccessModal: undefined,
-  version: LATEST_METADATA_VERSION,
-}
-
-// IF CHANGING ANY DEFAULT STATE, MUST INCREMENT REDUX VERSION ABOVE (`REDUX_STORE_V2_PROJECT_VERSION`)
-const defaultProjectState: ProjectState = {
-  projectMetadata: { ...defaultProjectMetadataState },
-  fundingCycleData: { ...defaultFundingCycleData },
-  fundingCycleMetadata: { ...defaultFundingCycleMetadata },
-  fundAccessConstraints: [],
-  payoutGroupedSplits: EMPTY_PAYOUT_GROUPED_SPLITS,
-  reservedTokensGroupedSplits: EMPTY_RESERVED_TOKENS_GROUPED_SPLITS,
-  nftRewards: {
-    rewardTiers: [],
-    CIDs: undefined,
-    collectionMetadata: EMPTY_NFT_COLLECTION_METADATA,
-    postPayModal: undefined,
-    flags: DEFAULT_NFT_FLAGS,
-    contractVersion: JB721_DELEGATE_V1,
-    governanceType: JB721GovernanceType.NONE,
-  },
-  mustStartAtOrAfter: DEFAULT_MUST_START_AT_OR_AFTER,
-  inputProjectOwner: undefined,
-}
-
-export const defaultReduxState: ReduxState = {
-  version: REDUX_STORE_V2_PROJECT_VERSION,
-  ...defaultProjectState,
-  ...defaultCreateState,
-}
+import {
+  DEFAULT_REDUX_STATE,
+  EMPTY_PAYOUT_GROUPED_SPLITS,
+  EMPTY_RESERVED_TOKENS_GROUPED_SPLITS,
+} from './defaultState'
+import { NftRewardsData, ReduxState } from './types'
 
 const editingV2ProjectSlice = createSlice({
   name: 'editingV2Project',
-  initialState: defaultReduxState,
+  initialState: DEFAULT_REDUX_STATE,
   reducers: {
     setState: (_, action: PayloadAction<ReduxState>) => {
       return action.payload
     },
-    resetState: () => defaultReduxState,
+    resetState: () => DEFAULT_REDUX_STATE,
     setName: (state, action: PayloadAction<string>) => {
       state.projectMetadata.name = action.payload
     },
