+++ src/redux/slices/editingV2Project/index.ts
@@ -1,0 +1,5 @@
+export * from './editingV2Project'
+export * from './types'
+export * from './defaultState'
+
+export { default } from './editingV2Project'
