+++ src/utils/v2v3/math.ts
@@ -1,18 +1,9 @@
 import { BigNumber } from '@ethersproject/bignumber'
 import * as constants from '@ethersproject/constants'
-import { invertPermyriad } from 'utils/bigNumbers'
-import {
-  formattedNum,
-  fromWad,
-  percentToPermyriad,
-} from 'utils/format/formatNumber'
-import { WeightFunction } from 'utils/math'
-
 import {
   MaxUint232,
   MaxUint88,
   ONE_BILLION,
-  ONE_MILLION,
   TEN_THOUSAND,
 } from 'constants/numbers'
 import { parseEther } from 'ethers/lib/utils'
@@ -21,6 +12,13 @@ import {
   V2V3FundingCycleData,
   V2V3FundingCycleMetadata,
 } from 'models/v2v3/fundingCycle'
+import { invertPermyriad } from 'utils/bigNumbers'
+import {
+  formattedNum,
+  fromWad,
+  percentToPermyriad,
+} from 'utils/format/formatNumber'
+import { WeightFunction } from 'utils/math'
 
 export const MAX_RESERVED_RATE = TEN_THOUSAND
 export const MAX_REDEMPTION_RATE = TEN_THOUSAND
@@ -28,7 +26,6 @@ export const MAX_DISCOUNT_RATE = ONE_BILLION
 export const SPLITS_TOTAL_PERCENT = ONE_BILLION
 export const MAX_DISTRIBUTION_LIMIT = MaxUint232
 
-export const DEFAULT_MINT_RATE = ONE_MILLION
 export const MAX_MINT_RATE = Math.floor(MaxUint88 / 10 ** 18)
 
 const MAX_FEE = ONE_BILLION
